+++ config/enquire.php
@@ -3,6 +3,10 @@
 return [
     'fields' => [
         'checkbox'      => 'enquire::partials.checkbox',
+        'checkboxes'    => 'enquire::partials.checkboxes',
+        'dropdown'      => 'enquire::partials.dropdown',
+        'email'         => 'enquire::partials.email',
+        'file_upload'   => 'enquire::partials.file',
         'phone'         => 'enquire::partials.phone',
         'radio_buttons' => 'enquire::partials.radios',
         'text'          => 'enquire::partials.text',
@@ -13,6 +17,7 @@ return [
     'fields_with_options' => [
         'dropdown',
         'radio_buttons',
+        'checkboxes',
     ],
 
     'from_address' => '',